Clean up directories and update gender distribution to read real data
